requirements.txt
# Web scraping
scrapy==2.11.0
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.0

# Data processing
pandas==2.1.3
numpy==1.26.2
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2

# Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
elasticsearch==8.11.0
redis==5.0.1
alembic==1.12.1

# ML/AI
torch==2.1.1
scikit-learn==1.3.2
sentence-transformers==2.2.2
openai==1.3.0

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
pydantic==2.5.2
python-multipart==0.0.6

# Configuration & Utils
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
schedule==1.2.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0